Fix badge, yet again
'3D' badge is not always added.
$badge is declared at the top of the function, and we used
to add it to all $elements at once. This worked well.

We recently added a promise that doesn't resolve until an
individual image resolves, and the badge is not to be added
until after that promise is resolved. This also mostly works.

The problem is with $badge, though. It is declared up top and
treated as 1 node. So when it's later added to multiple images
(e.g. main image + revision), it is the same node that is
moved around, instead of multiple instances of that badge.

So the easy fix is to hold off creating that node until inside
the promise, where it's then guaranteed to be unique per image.

-

I made the same change in mmv.3d.head.js for the "view" &
"download" controls that are added. The problem was not
noticeable there because it is only added to the 1 main thumb.
But better fix it here properly as well :)

Bug: T187180
Change-Id: If7fb6db9858606749b6d4e19a5ffac4384c12374