Avoid height jump (FOUC) when showing 'loading' placeholder
Put a min-height on the placeholder equal to the
height of the image.

Also put the placeholder insider the wrapper and
toggle the image instead, to avoid inline-block
height issues.

Generate $placeholder template just once.

And wait just a little bit before hiding the image, to
minimize FOUC when the image is already available.

Bug: T187327
Change-Id: Ie650bb8a429ce3dc1298ab436f28e63805a81eb7