Hide 'Patent permisisons' radio buttons by default in Special:Upload
JavaScript will show/hide these options depending on whether the upload
is a relevant (3D) file or not, so this is just the default for users
without JS, who - up until now - would always see the 'Patent
permissions', often get confused and select something that is not
relevant to the non-3D file they're uploading.

Of course, this change will make it so that people uploading 3D files
without JavaScript will not have the option to select the appropriate
patent permissions. However, the volume of those will be much lower,
and there is a category to track such uploads:
Category:Files_with_no_machine-readable_patent

Bug: T188258
Change-Id: I110c24dba626acc3f3fa7af7938198a17dc99612