Try to consistently call this extension 3D instead of 3d.
The mediawiki.org page uses 3D. Conventional english usage is
to usually capitalize the D when talking about 3-D. The git
repo capitalizes the D.

Additionally, having the mediawiki ext url in extension.json not be
the same as the actual page on mediawiki.org confuses some
templates.

For now, intentionally not changing anything that would have
a code impact.

Change-Id: Ife34dd633b8ce489162b7b85663a60f22fe5f0f9